requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Voice & Speech
twilio==8.10.0
deepgram-sdk==3.0.0
elevenlabs==0.2.27
soundfile==0.12.1
pydub==0.25.1

# AI & NLP
anthropic==0.7.8
openai==1.3.7
langchain==0.0.340
langchain-anthropic==0.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
websockets==12.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3
